Eliminated some "comparison between signed and unsigned integer expressions" warnings
git-svn-id: http://usg.lofar.org/svn/code/trunk@1347 e3a606d5-b720-0410-b6ad-cb58f335abf8

+++ src/CR-Tools/implement/IO/ApplyInstrumentEffects.cc
@@ -88,7 +88,7 @@ namespace CR { // Namespace CR -- begin
 
       nAnts = input.ncolumn();
       nFreq = input.nrow();
-      if ((nAnts != AntIDs_p.nelements()) || (nFreq != frequency_p.nelements())){
+      if ((nAnts != (int)AntIDs_p.nelements()) || (nFreq != (int)frequency_p.nelements())){
 	cerr << "ApplyInstrumentEffects::ApplyEffects: " << 
 	  "Shape of input matrix does not match AntIDs or frequencies!!" << endl;
 	return Matrix<DComplex>();
